requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==4
annotated-types==0.5.0
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
decorator==5.1.1
dnspython==2.4.2
elevenlabs==0.2.26
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.35
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
loguru==0.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
openai==0.28.0
openpyxl==3.1.2
packaging==23.1
pandas-stubs==2.0.3.230814
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pinecone-client==2.2.2
prompt-toolkit==3.0.39
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.1b0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.10.2
rsa==4.9
semver==3.0.1
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.25.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
types-pytz==2023.3.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
validators==0.22.0
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
youtube-transcript-api==0.6.1
zipp==3.16.2
